requirements.txt
python-telegram-bot
peewee
psycopg2-binary